renamed directory for naming consistency
